fix: promise error and add translation in form
