+++ src/features/auth/api/instanceApi.ts
@@ -32,14 +32,10 @@ instance.interceptors.response.use(
                 return axios(originalRequest)
               }
             })
-            .catch((error) => {
-              console.log(error)
-            })
+            .catch(() => {})
         }
       }
     }
-
-    return Promise.reject(error)
   }
 )
